move API `memfrs_enum_proc_list` out of memfrs.c
